+++ test/suites/BrowserFactoryTest.php
@@ -6,6 +6,7 @@
 namespace HeadlessChromium\Test;
 
 use HeadlessChromium\BrowserFactory;
+use HeadlessChromium\Communication\Target;
 
 /**
  * @covers \HeadlessChromium\BrowserFactory
@@ -13,6 +14,14 @@ use HeadlessChromium\BrowserFactory;
  */
 class BrowserFactoryTest extends BaseTestCase
 {
+    public function testBrowserFactory()
+    {
+        $factory = new BrowserFactory();
+
+        $browser = $factory->createBrowser();
+
+        $this->assertRegExp('#^ws://#', $browser->getSocketUri());
+    }
 
     public function testWindowSizeOption()
     {
@@ -44,12 +53,26 @@ class BrowserFactoryTest extends BaseTestCase
         $this->assertEquals('foo bar baz', $response);
     }
 
-    public function testBrowserFactory()
+    public function testConnectToBrowser()
     {
+        // create a browser
         $factory = new BrowserFactory();
-
         $browser = $factory->createBrowser();
 
-        $this->assertRegExp('#^ws://#', $browser->getSocketUri());
+        // TODO test existing pages propagation
+
+        // create a new connectionn to the existing browser
+        $browser2 = BrowserFactory::connectToBrowser($browser->getSocketUri());
+
+        // create a page on the first browser after 2d connection
+        $page2 = $browser->createPage();
+        $page2TargetId = $page2->getSession()->getTargetId();
+
+        // update 2d browser
+        $browser2->getConnection()->readData();
+
+        // make sure 2nd browser received the new page
+        $target = $browser2->getTarget($page2TargetId);
+        $this->assertInstanceOf(Target::class, $target);
     }
 }